I made two changes to these drivers. EcalRawConverterDriver: I just remove an old comment. Eval RawConverter: in the sumADC function, called when we convert from MODE1 to energy, the return value was a short. a short is too small! We need an int. I am sure I already did this modification, but now is back to short... please do not change it again.

+++ ecal-recon/src/main/java/org/hps/recon/ecal/EcalRawConverter.java
@@ -154,10 +154,10 @@ public class EcalRawConverter {
     /*
      * This should probably be deprecated.  It just integrates the entire window.
      */
-    public short sumADC(RawTrackerHit hit) {
+    public int sumADC(RawTrackerHit hit) {
         EcalChannelConstants channelData = findChannel(hit.getCellID());
         double pedestal = channelData.getCalibration().getPedestal();
-        short sum = 0;
+        int sum = 0;
         short samples[] = hit.getADCValues();
         for (int isample = 0; isample < samples.length; ++isample) {
             sum += (samples[isample] - pedestal);
